+++ packages/kokoas-client/src/pages/projRegister/sections/CustInfo/CustInfo.tsx
@@ -5,7 +5,7 @@ import { PageSubTitle } from '../../../../components/ui/labels/';
 import EditIcon from '@mui/icons-material/Edit';
 import {  useNavigate } from 'react-router-dom';
 import { useFormikContext } from 'formik';
-import { TypeOfForm, getFieldName } from '../../form';
+import { TypeOfForm } from '../../form';
 import { CustomerInstance } from '../../../customer/register/form';
 import { pages } from '../../../Router';
 import { EmptyBox } from '../../../../components/ui/information/EmptyBox';
@@ -14,24 +14,19 @@ import { Column1 } from './Column1';
 import { Column2 } from './Column2';
 import { LabeledInfo } from '../../../../components/ui/typographies';
 import { AGLabels } from 'types';
-import { SearchCustGroup } from 'kokoas-client/src/components/ui/textfield/SearchCustGroup';
 import { useCustGroupById } from 'kokoas-client/src/hooksQuery';
-import { useConfirmDialog } from 'kokoas-client/src/hooks';
-import { ContentWarning } from 'kokoas-client/src/components/ui/dialogs/ContentWarning';
-
-
+import { RecordSelect } from '../RecordSelect/RecordSelect';
 
 export const CustInfo = () => {
 
-  const { status, values, dirty } = useFormikContext<TypeOfForm>();
+  const { status, values } = useFormikContext<TypeOfForm>();
   const navigate = useNavigate();
-  const { setDialogState } = useConfirmDialog();
 
   const isReadOnly = (status as TFormStatus ) === 'disabled';
 
   const {
     custGroupId,
-    recordId,
+    projId,
   } = values;
 
   const { data: custGroupRecord } = useCustGroupById(custGroupId ?? '');
@@ -57,13 +52,6 @@ export const CustInfo = () => {
     phone2, phone2Rel,
   } = JSON.parse(dump?.value || 'null') as CustomerInstance ?? {};
 
-  /* 工事名を生成する */
-  /*   useEffect(()=>{
-    setValues(prev => ({
-      ...prev,
-      projName: `${prev.custName ?? '--'}様邸 ${projTypeName ?? '--'}`,
-    }));
-  }, [projTypeName, setValues]); */
 
   const refactoredAgents = custGroupRecord?.agents
     .value
@@ -78,48 +66,10 @@ export const CustInfo = () => {
     }, [] as Array<ComponentProps<typeof LabeledInfo> & { key: string }>) ?? [];
 
 
-  const navigateToCustGroup = (newCustGroupId: string) => {
-    navigate(`${pages.projEdit}?${generateParams({
-      custGroupId: newCustGroupId,
-    })}`);
-  };
 
   return (
     <>
-      <Grid item xs={12} md={4} >
-        <SearchCustGroup
-          value={custGroupId ? {
-            id: custGroupId,
-            name: customerName?.value || '',
-            record: custGroupRecord,
-          } : undefined}
-          onChange={(_, val) => {
-            const { id: newCustGroupId } = val || {};
-            if (newCustGroupId ) {
-              if (dirty) {
-                setDialogState({
-                  open: true,
-                  title: '動作確認',
-                  content: <ContentWarning content={'顧客を変更すると工事情報がリセットされます。'} />,
-                  withNo: true,
-                  withYes: true,
-                  yesText: 'OK',
-                  noText: 'キャンセル',
-                  handleYes: () => {
-                    navigateToCustGroup(newCustGroupId);
-                  },
-                });
-              } else {
-                navigateToCustGroup(newCustGroupId);
-              }
-            }
-          }}
-          inputProps={{
-            label: 'お客検索',
-            name: getFieldName('custGroupId'),
-          }}
-        />
-      </Grid>
+      <RecordSelect />
       <PageSubTitle label="顧客情報" />
       <Grid item xs={12}>
 
@@ -158,7 +108,7 @@ export const CustInfo = () => {
                     startIcon={<EditIcon />}
                     onClick={()=>navigate(`${pages.custGroupEdit}?${generateParams({
                       custGroupId,
-                      projId: recordId,
+                      projId,
                     })}`)}
                     fullWidth
