+++ packages/kokoas-client/src/pages/projRegister/sections/RecordSelect/RecordSelect.tsx
@@ -1,0 +1,90 @@
+import { Grid } from '@mui/material';
+import { useFormikContext } from 'formik';
+import { ContentWarning } from 'kokoas-client/src/components/ui/dialogs/ContentWarning';
+import { ModeInfo } from 'kokoas-client/src/components/ui/information/ModeInfo';
+import { SearchCustGroup } from 'kokoas-client/src/components/ui/textfield/SearchCustGroup';
+import { generateParams } from 'kokoas-client/src/helpers/url';
+import { useConfirmDialog } from 'kokoas-client/src/hooks';
+import { pages } from 'kokoas-client/src/pages/Router';
+import { useNavigate } from 'react-router-dom';
+import { getFieldName, TypeOfForm } from '../../form';
+
+export const RecordSelect = () => {
+  const {
+    values,
+    dirty,
+  } = useFormikContext<TypeOfForm>();
+
+  const {
+    projId,
+    custGroupId,
+    custName,
+    createdDate,
+  } = values;
+
+  const { setDialogState } = useConfirmDialog();
+  const navigate = useNavigate();
+  const navigateToCustGroup = (newCustGroupId: string) => {
+    navigate(`${pages.projEdit}?${generateParams({
+      custGroupId: newCustGroupId,
+    })}`);
+  };
+
+  return (
+    <Grid
+      container
+      item
+      xs={12}
+      spacing={2}
+    >
+      <Grid item
+        xs={12}
+        md={4}
+      >
+        <SearchCustGroup
+          fullWidth
+          value={custGroupId ? {
+            id: custGroupId,
+            name: custName,
+          } : undefined}
+          onChange={(_, val) => {
+            const { id: newCustGroupId } = val || {};
+            if (newCustGroupId ) {
+              if (dirty) {
+                setDialogState({
+                  open: true,
+                  title: '動作確認',
+                  content: <ContentWarning content={'顧客を変更すると工事情報がリセットされます。'} />,
+                  withNo: true,
+                  withYes: true,
+                  yesText: 'OK',
+                  noText: 'キャンセル',
+                  handleYes: () => {
+                    navigateToCustGroup(newCustGroupId);
+                  },
+                });
+              } else {
+                navigateToCustGroup(newCustGroupId);
+              }
+            }
+          }}
+          inputProps={{
+            label: 'お客検索',
+            name: getFieldName('custGroupId'),
+          }}
+        />
+      </Grid>
+      <Grid
+        item
+        xs={12}
+        md={4}
+      >
+        <ModeInfo
+          recordId={projId || ''}
+          dateStr={createdDate}
+        />
+      </Grid>
+
+    </Grid>
+  );
+};
